Consider all entries in undo-tree changesets
Each undo-tree node stores a list of undo entries delimited by undo
boundaries. This commit ensures that all such entries are considered and
not just the first one in each list.

Also, undo-tree canaries are an internal detail of that package, see
https://www.dr-qubit.org/Lost_undo-tree_history.html for an explanation,
so this commit removes the whole section dealing with those.

Resolves #12